(maint) Add ability to specify a ref to compare against
This change allows the GitHub action workflow to specify a specific
reference to be used when finding files to send to the language server.

This allows the the languagetool check to use less time by only compare
any new or changed files, instead of running all of the files in the
repository through the language server.

Additional work is needed for repositories after this change is merged.